Added methods to concrete builders, fixed file name, added temp iterator into ClassDataHandler
The concrete builders are utilizing the methods that were used to test grabbing info from the course booklet. (See Class.cs, the methods used to write to the Console). The iterator at the end of ClassDataHandler is supposed to mimic the separation of each courses info, however, it does not account for new lines (I'm pretty sure).